+++ Application.cpp
@@ -68,6 +68,7 @@ namespace Coda {
 
 			IF_ERROR_RETURN(1);
 			Coda::Frontend::Program program = parse(tokens);
+
 			IF_ERROR_RETURN(1);
 			interpret(program, env);
 		}